pm/drivers: fix MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012
If a composite expression is used as one operand
of an operator in which the usual arithmetic conversions
are performed then the other operand shall not have wider
essential type.

Fixed by typecasting to the required datatype.

Fixes: SYSFW-6414

+++ pm/drivers/clock/clk_mux.c
@@ -105,9 +105,9 @@ static sbool clk_mux_set_parent(struct clk *clkp, u8 new_parent)
 			ret = SFALSE;
 		} else {
 			pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-				 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+				 (((u32) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 				  TRACE_PM_VAL_CLOCK_VAL_MASK) |
-				 ((clk_id(clkp) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
+				 (((u32)clk_id(clkp) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 				  TRACE_PM_VAL_CLOCK_ID_MASK));
 		}
 	}
